Add files via upload
Tried to build a blockchain protocol by following a book, but it's not really the work of a single person 😅.  To this point, the protocol can 
Let you add a new transaction, 
Mine a new block, and also gives you a reward for mining a block. 
Throws a tantrum if someone tries to manipulate it.
These are the key features that I was able to bring in it. At the time of the writing am learning about cryptography, and I would try to bring all the things one by one in this project.